+++ examples/2_curved_slicing/ex2_curved_slicing_advanced.py
@@ -3,15 +3,15 @@ from compas.datastructures import Mesh
 import logging
 import compas_slicer.utilities as utils
 from compas_slicer.slicers import InterpolationSlicer
-from compas_slicer.post_processing import simplify_paths_rdp
+from compas_slicer.post_processing import simplify_paths_rdp_igl
 from compas_slicer.pre_processing import InterpolationSlicingPreprocessor
 from compas_slicer.pre_processing import create_mesh_boundary_attributes
 from compas_slicer.print_organization import InterpolationPrintOrganizer
 from compas_slicer.print_organization import set_extruder_toggle
-from compas_slicer.print_organization import add_safety_printpoints
-from compas_slicer.print_organization import smooth_printpoints_up_vectors
-from compas_slicer.post_processing import generate_brim
-import time
+from compas_slicer.print_organization import add_safety_printpoints, set_wait_time_on_sharp_corners
+from compas_slicer.print_organization import smooth_printpoints_up_vectors, set_blend_radius
+from compas_slicer.post_processing import generate_brim, seams_smooth
+import time, math
 
 logger = logging.getLogger('logger')
 logging.basicConfig(format='%(levelname)s - %(message)s', level=logging.INFO)
@@ -29,12 +29,13 @@ PRINT_ORGANIZER = True
 def main():
     start_time = time.time()
 
+    avg_layer_height = 4.0
     parameters = {
-        'target_LOW_smooth_union': [True, 7],  # boolean, blend_radius
-        'target_HIGH_smooth_union': [True, 7],  # boolean, blend_radius
-        'avg_layer_height': 5.0,  # controls number of curves that will be generated
-        'min_layer_height': 0.1,
-        'max_layer_height': 50.0  # 2.0
+        'avg_layer_height': avg_layer_height,  # controls number of curves that will be generated
+        'min_layer_height': 0.2,
+        'max_layer_height': 4.0,
+        'uneven_upper_targets_offset': 0,
+        'target_HIGH_smooth_union': [True, [25.0]],  # on/off, blend radius
     }
 
     ### --- Load initial_mesh
@@ -64,6 +65,7 @@ def main():
     #########################################
     # --- slicing
     if SLICER:
+
         slicers = []
         filenames = utils.get_all_files_with_name('split_mesh_', '.json', OUTPUT_PATH)
         split_meshes = [Mesh.from_json(os.path.join(OUTPUT_PATH, filename)) for filename in filenames]
@@ -83,7 +85,8 @@ def main():
             if i == 0:
                 generate_brim(slicer, layer_width=3.0, number_of_brim_offsets=5)
 
-            simplify_paths_rdp(slicer, threshold=1.0)
+            seams_smooth(slicer, smooth_distance=0.1)
+            simplify_paths_rdp_igl(slicer, threshold=0.25)
             utils.save_to_json(slicer.to_data(), OUTPUT_PATH, 'curved_slicer_%d.json' % i)
             slicers.append(slicer)
 
@@ -95,14 +98,15 @@ def main():
         filenames = utils.get_all_files_with_name('curved_slicer_', '.json', OUTPUT_PATH)
         slicers = [InterpolationSlicer.from_data(utils.load_from_json(OUTPUT_PATH, filename)) for filename in filenames]
         for i, slicer in enumerate(slicers):
-
             print_organizer = InterpolationPrintOrganizer(slicer, parameters, DATA_PATH)
             print_organizer.create_printpoints()
             set_extruder_toggle(print_organizer, slicer)
+            set_blend_radius(print_organizer)
             add_safety_printpoints(print_organizer, z_hop=10.0)
             smooth_printpoints_up_vectors(print_organizer, strength=0.5, iterations=10)
+            set_wait_time_on_sharp_corners(print_organizer, threshold=0.5 * math.pi, wait_time=0.2)
 
-            ### --- Save printpoints dictionary to json file
+            # --- Save printpoints dictionary to json file
             printpoints_data = print_organizer.output_printpoints_dict()
             utils.save_to_json(printpoints_data, OUTPUT_PATH, 'out_printpoints_%d.json' % i)