+++ examples/2_curved_slicing/ex2_curved_slicing_basic.py
@@ -3,7 +3,7 @@ from compas.datastructures import Mesh
 import logging
 import compas_slicer.utilities as utils
 from compas_slicer.slicers import InterpolationSlicer
-from compas_slicer.post_processing import simplify_paths_rdp
+from compas_slicer.post_processing import simplify_paths_rdp_igl
 from compas_slicer.pre_processing import InterpolationSlicingPreprocessor
 from compas_slicer.print_organization import set_extruder_toggle, set_linear_velocity_by_range
 from compas_slicer.print_organization import add_safety_printpoints
@@ -38,8 +38,8 @@ def main():
 
     parameters = {
         'avg_layer_height': avg_layer_height,  # controls number of curves that will be generated
-        'min_layer_height': 0.1,
-        'max_layer_height': 50.0  # 2.0,
+        'min_layer_height': 0.3,
+        'max_layer_height': 5.0  # 2.0,
     }
 
     preprocessor = InterpolationSlicingPreprocessor(mesh, parameters, DATA_PATH)
@@ -55,11 +55,11 @@ def main():
     generate_brim(slicer, layer_width=3.0, number_of_brim_offsets=5)
     seams_smooth(slicer, smooth_distance=10)
 
-    simplify_paths_rdp(slicer, threshold=1.0)
+    simplify_paths_rdp_igl(slicer, threshold=0.5)
     slicer.printout_info()
     utils.save_to_json(slicer.to_data(), OUTPUT_PATH, 'curved_slicer.json')
 
-    # ### --- Print organizer
+    # --- Print organizer
     print_organizer = InterpolationPrintOrganizer(slicer, parameters, DATA_PATH)
     print_organizer.create_printpoints()