+++ examples/4_gcode_generation/example_4_gcode.py
@@ -4,7 +4,7 @@ import compas_slicer.utilities as utils
 from compas_slicer.pre_processing import move_mesh_to_point
 from compas_slicer.slicers import PlanarSlicer
 from compas_slicer.post_processing import generate_brim
-from compas_slicer.post_processing import simplify_paths_rdp
+from compas_slicer.post_processing import simplify_paths_rdp_igl
 from compas_slicer.post_processing import seams_smooth
 from compas_slicer.print_organization import PlanarPrintOrganizer
 from compas_slicer.print_organization import set_extruder_toggle
@@ -37,7 +37,7 @@ def main():
     slicer = PlanarSlicer(compas_mesh, slicer_type="cgal", layer_height=4.5)
     slicer.slice_model()
     generate_brim(slicer, layer_width=3.0, number_of_brim_offsets=4)
-    simplify_paths_rdp(slicer, threshold=0.6)
+    simplify_paths_rdp_igl(slicer, threshold=0.6)
     seams_smooth(slicer, smooth_distance=10)
     slicer.printout_info()
     save_to_json(slicer.to_data(), OUTPUT_DIR, 'slicer_data.json')
@@ -50,7 +50,7 @@ def main():
     print_organizer.printout_info()
 
     # create and output gcode
-    gcode_parameters = {}
+    gcode_parameters = {}  # leave all to default
     gcode_text = print_organizer.output_gcode(gcode_parameters)
     utils.save_to_text_file(gcode_text, OUTPUT_DIR, 'my_gcode.gcode')