+++ src/compas_slicer/print_organization/print_organization_utilities/blend_radius.py
@@ -26,14 +26,18 @@ def set_blend_radius(print_organizer, d_fillet=10, buffer=0.3):
         path_key = 'path_%d' % j
         neighboring_items = print_organizer.get_printpoint_neighboring_items(layer_key, path_key, k)
 
-        radius = d_fillet
-        if neighboring_items[0]:
-            radius = min(radius, norm_vector(Vector.from_start_end(neighboring_items[0].pt, printpoint.pt)) * buffer)
+        if not printpoint.wait_time:
+            radius = d_fillet
+            if neighboring_items[0]:
+                radius = min(radius, norm_vector(Vector.from_start_end(neighboring_items[0].pt, printpoint.pt)) * buffer)
 
-        if neighboring_items[1]:
-            radius = min(radius, norm_vector(Vector.from_start_end(neighboring_items[1].pt, printpoint.pt)) * buffer)
+            if neighboring_items[1]:
+                radius = min(radius, norm_vector(Vector.from_start_end(neighboring_items[1].pt, printpoint.pt)) * buffer)
 
-        radius = round(radius, 5)
+            radius = round(radius, 5)
+
+        else:
+            radius = 0.0  # 0.0 blend radius for points where the robot will pause and wait
 
         printpoint.blend_radius = radius