+++ src/compas_slicer/print_organization/print_organization_utilities/gcode.py
@@ -65,7 +65,7 @@ def create_gcode_text(print_organizer, parameters):
     # ######################################################################
     # gcode header
     gcode += ";Gcode with compas_slicer " + n_l
-    gcode += ";Ioana Mitropolou <mitropoulou@arch.ethz.ch> @ioanna21" + n_l
+    gcode += ";Ioana Mitropolou <mitropoulou@arch.ethz.ch> @ioannaMitropoulou" + n_l
     gcode += ";Joris Burger     <burger@arch.ethz.ch>      @joburger" + n_l
     gcode += ";Andrei Jipa      <jipa@arch.ethz.ch>         @stratocaster" + n_l
     gcode += ";MIT License" + n_l