+++ src/compas_slicer/print_organization/print_organization_utilities/linear_velocity.py
@@ -11,19 +11,21 @@ __all__ = ['set_linear_velocity_constant',
 
 
 def set_linear_velocity_constant(print_organizer, v=25.0):
-    """ Sets the linear velocity parameter of the printpoints depending on the selected type.
+    """Sets the linear velocity parameter of the printpoints depending on the selected type.
 
     Parameters
     ----------
     print_organizer: :class:`compas_slicer.print_organization.BasePrintOrganizer`
-    v:  float. Velocity value (in mm/s) to set for printpoints. Defaults to 25 mm/s. """
+    v:  float. Velocity value (in mm/s) to set for printpoints. Defaults to 25 mm/s.
+    """
+
     logger.info("Setting constant linear velocity")
     for printpoint in print_organizer.printpoints_iterator():
         printpoint.velocity = v
 
 
 def set_linear_velocity_per_layer(print_organizer, per_layer_velocities):
-    """ Sets the linear velocity parameter of the printpoints depending on the selected type.
+    """Sets the linear velocity parameter of the printpoints depending on the selected type.
 
     Parameters
     ----------
@@ -31,6 +33,7 @@ def set_linear_velocity_per_layer(print_organizer, per_layer_velocities):
     per_layer_velocities: list
         A list of velocities (floats) with equal length to the number of layers.
     """
+
     logger.info("Setting per-layer linear velocity")
     assert len(per_layer_velocities) == print_organizer.number_of_layers, 'Wrong number of velocity values. You need \
         to provide one velocity value per layer, on the "per_layer_velocities" list.'
@@ -40,8 +43,7 @@ def set_linear_velocity_per_layer(print_organizer, per_layer_velocities):
 
 def set_linear_velocity_by_range(print_organizer, param_func, parameter_range, velocity_range,
                                  bound_remapping=True):
-    """
-    Sets the linear velocity parameter of the printpoints depending on the selected type.
+    """Sets the linear velocity parameter of the printpoints depending on the selected type.
 
     Parameters
     ----------
@@ -55,6 +57,7 @@ def set_linear_velocity_by_range(print_organizer, param_func, parameter_range, v
     bound_remapping: bool
         If True, the remapping is bound in the domain velocity_range, else it is unbound.
     """
+
     logger.info("Setting linear velocity based on parameter range")
     for printpoint in print_organizer.printpoints_iterator():
         param = param_func(printpoint)
@@ -67,7 +70,8 @@ def set_linear_velocity_by_range(print_organizer, param_func, parameter_range, v
 
 
 def set_linear_velocity_by_overhang(print_organizer, overhang_range, velocity_range, bound_remapping=True):
-    """
+    """Set velocity by overhang by using set_linear_velocity_by_range.
+
     An example function for how to use the 'set_linear_velocity_by_range'. In this case the parameter that controls the
     velocity is the overhang, measured as a dot product with the horizontal direction.
 
@@ -80,6 +84,7 @@ def set_linear_velocity_by_overhang(print_organizer, overhang_range, velocity_ra
     velocity_range: tuple
     bound_remapping: bool
     """
+
     def param_func(ppt): return dot_vectors(ppt.mesh_normal, Vector(0.0, 0.0, 1.0))
     # returns values from 0.0 (no overhang) to 1.0 (horizontal overhang)
     set_linear_velocity_by_range(print_organizer, param_func, overhang_range, velocity_range, bound_remapping)